+++ test/integration/ssl-56/ssl_spec.rb
@@ -1,0 +1,3 @@
+require_relative '../helpers/server_spec.rb'
+
+server_test('5.6', 'ssl')